requirements.txt
Flask==2.3.3
pandas==2.0.3
numpy==1.24.4
yfinance==0.2.28
plotly==5.18.0
dash==2.13.0
flask-wtf==1.2.1
python-dotenv==1.0.0
gunicorn==21.2.0
backtesting==0.3.3 